requirements.txt
fastapi[all]
requests
faiss-cpu
sentence-transformers
uvicorn
httpx
jinja2
psycopg2
python-dotenv
redis
pymongo